+++ pytorch_lightning/trainer/connectors/accelerator_connector.py
@@ -769,10 +769,6 @@ class AcceleratorConnector:
         return self._parallel_devices
 
     @property
-    def num_nodes(self) -> int:
-        return self._num_nodes_flag
-
-    @property
     def num_processes(self) -> int:
         return self.devices if self.devices is not None else 1